v1.2.2 to replace `double` with `float`
### Releases v1.2.2

1. Use `float` for `DutyCycle` and `Freq`, `uint32_t` for `period`. 
2. Optimize code by not calculation in ISR

+++ examples/ISR_Changing_PWM/ISR_Changing_PWM.ino
@@ -69,19 +69,19 @@ bool IRAM_ATTR TimerHandler(void * timerNo)
 uint32_t PWM_Pin    = LED_BUILTIN;
 
 // You can assign any interval for any timer here, in Hz
-double PWM_Freq1   = 1.0f;
+float PWM_Freq1   = 1.0f;
 // You can assign any interval for any timer here, in Hz
-double PWM_Freq2   = 2.0f;
+float PWM_Freq2   = 2.0f;
 
 // You can assign any interval for any timer here, in microseconds
-double PWM_Period1 = 1000000.0 / PWM_Freq1;
+uint32_t PWM_Period1 = 1000000 / PWM_Freq1;
 // You can assign any interval for any timer here, in microseconds
-double PWM_Period2 = 1000000.0 / PWM_Freq2;
+uint32_t PWM_Period2 = 1000000 / PWM_Freq2;
 
 // You can assign any duty_cycle for any PWM here, from 0-100
-double PWM_DutyCycle1  = 50.0;
+float PWM_DutyCycle1  = 50.0;
 // You can assign any duty_cycle for any PWM here, from 0-100
-double PWM_DutyCycle2  = 90.0;
+float PWM_DutyCycle2  = 90.0;
 
 // Channel number used to identify associated channel
 int channelNum;
@@ -135,7 +135,7 @@ void loop()
 
   delay(10000);
 
-  ISR_PWM.deleteChannel((unsigned) channelNum);
+  ISR_PWM.deleteChannel((uint8_t) channelNum);
 
   Serial.print(F("Using PWM Freq = ")); Serial.print(PWM_Freq2); Serial.print(F(", PWM DutyCycle = ")); Serial.println(PWM_DutyCycle2);
 
@@ -156,5 +156,5 @@ void loop()
 
   delay(10000);
 
-  ISR_PWM.deleteChannel((unsigned) channelNum);
+  ISR_PWM.deleteChannel((uint8_t) channelNum);
 }
